Examples/public-docs/examples/cloud/requirements.txt
bithuman>=1.6.0
livekit-agents[openai,bithuman,silero]~=1.3
python-dotenv~=1.1
Pillow>=10.0
